Link game card with game list
